get styling working on users
