+++ src/components/Home/Home.tsx
@@ -20,77 +20,7 @@ import * as history from '../helpers/volumeHistoryHelper';
 
 const Home = (): JSX.Element => {
 
-  const { sessions, volumes } = useAppSelector((state) => state);
-  const userData = sessions;
-  const { arrayOfVolumeNames } = volumes;
 
-  const {
-    refreshRunning,
-    refreshStopped,
-    refreshImages,
-    writeToDb,
-    networkContainers,
-    // below function never called
-    // setDbSessionTimeZone,
-    getAllDockerVolumes,
-    handlePruneClick,
-    getVolumeContainers,
-  } = useHelper();
-
-  // Deconstructs dispatch functions from custom hook
-  const { updateUser, getVolumeContainerList } =
-    useSurvey();
-
-  useEffect(() => {
-    refreshRunning();
-    refreshStopped();
-    refreshImages();
-    writeToDb();
-    networkContainers();
-    getAllDockerVolumes();
-    setAdminToken();
-  }, []);
-
-  // Changes in arrayOfVolumeNames will run history.volumeByName
-  useEffect(() => {
-    history.volumeByName(
-      getVolumeContainers,
-      arrayOfVolumeNames,
-      getVolumeContainerList
-    );
-  }, [arrayOfVolumeNames]);
-
-  // Refresh runningList, stoppedList, and imageList every 5-seconds to ensure GUI accurately depicts local Docker environment
-  useEffect(() => {
-    const interval = setInterval(() => {
-      refreshRunning();
-      refreshStopped();
-      refreshImages();
-    }, 5000);
-    return () => clearInterval(interval);
-  }, []);
-
-  // Pertains to sysAdmin only
-  // TODO: double check what this is doing and name appropriately
-  const setAdminToken = async (): Promise<void> => {
-    try {
-      const response = await fetch('/api/admin', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          token: userData.token,
-          username: userData.username,
-        }),
-      });
-      const parsedData = await response.json();
-
-      updateUser(parsedData);
-    } catch (e) {
-      console.log(e);
-    }
-  };
 
   return (
     <>
