+++ src/components/SharedLayout/SharedLayout.module.scss
@@ -3,12 +3,28 @@
 }
 
 .navButton {
-
+  color: color(light, base);
+  padding: 0.5rem 1rem;
 }
 
+.active {
+  color: color(light, base);
+  padding: 0.5rem 1rem;
+  font-weight: 700;
+  text-decoration: underline;
+}
+
+
 .navBar {
-  display: flex;
-  margin-top: 1rem;
+  background-color: color(background, darker);
+  position: fixed;
+  z-index: 9999;
+  top: 0px;
+  left: 0px;
+  right: 0px;
+  padding: 0 1rem;
+  // transition: opacity 250ms;
+  // will-change: transform;
   li {
     list-style: none;
     margin: 1 1rem;
@@ -16,12 +32,17 @@
       text-decoration: none;
       color: color(white, tan);
       font-size: 1rem;
-      font-weight: 600;
-      &:hover {
-        color: color(primary, base);
-      }
     }
   }
+}
+
+.innerNav {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  width: 100%;
+  height: 10vh;
+
 }
 
 .footer {
