--- src/components/Users/Users.modules.scss
@@ -1,14 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  th {
-    color: red;
-  }
-  width: 70%;
-  margin: 0 15%;
-}
-
-h2 {
-  color: red;
-}
