+++ src/components/global.module.scss
@@ -7,7 +7,8 @@
   margin-bottom: 1rem;
   font-size: 1.2rem;
   color: color(white, tan);
-  background-color: color(darks, darker);
+  background-color: color(light, light-gray);
+  box-shadow: $shadow;
 }
 
 .radioForm{
